Optimizing for GIT submodules again.. :)
Turns out - the development story is actually a lot better using the GIT Submodules than what I thought an hour ago.. :)    No need for publishing to NPM and I get it all working seemlessly.

This particular push takes a dependency on the pipeline as a package through the Module folder. That way we don't have to explicitly have all the dev dependencies - easier!